+++ packages/core/theme/src/utils/variants.ts
@@ -9,13 +9,13 @@ const solid = {
 };
 
 const outline = {
-  primary: "border-brand text-brand-800 font-semibold",
-  danger: "border-danger text-danger-800 font-semibold",
-  warning: "border-warning text-warning-800 font-semibold",
-  success: "border-success text-success-800 font-semibold",
-  white: "border-white text-white font-semibold",
-  basic: "border-neutral text-neutral-800 font-semibold",
-  transparent: "border-transparent text-brand-800 font-semibold",
+  primary: "border-brand text-brand-800",
+  danger: "border-danger text-danger-800",
+  warning: "border-warning text-warning-800",
+  success: "border-success text-success-800",
+  white: "border-white text-white",
+  basic: "border-neutral text-neutral-800",
+  transparent: "border-transparent text-brand-800",
 };
 
 const ghost = {
@@ -24,8 +24,8 @@ const ghost = {
   warning: "text-warning-800",
   success: "text-success-800",
   white: "text-white",
-  basic: "text-white",
-  transparent: "text-white",
+  basic: "text-black",
+  transparent: "text-neutral",
 };
 
 const rounded = {