+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/card
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.0.1
+  - @jala-banyu/ripple@1.0.1
+
 ## 1.0.0
 
 ### Major Changes