+++ packages/components/code/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/code
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.0.1
+  - @jala-banyu/system-rsc@1.0.1
+
 ## 1.0.0
 
 ### Major Changes