+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/menu
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/theme@1.0.1
+  - @jala-banyu/divider@1.0.1
+
 ## 1.0.0
 
 ### Major Changes