+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jala-banyu/popover
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @jala-banyu/button@1.0.2
+  - @jala-banyu/theme@1.0.1
+
 ## 1.0.0
 
 ### Major Changes