+++ packages/core/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jala-banyu/theme
 
+## 1.0.1
+
+### Patch Changes
+
+- Update Button style
+
 ## 1.0.0
 
 ### Major Changes